Antora Integration (#48)
This commit introduces the integration of Antora documentation for libraries. With this enhancement, each library now has the capability to function as an Antora component within an Antora master project hosted in a separate repository (https://github.com/cppalliance/site-docs). This master project repository also contains additional components, such as the user guide, contributor guide, and a dedicated component for the review process.

In a subsequent phase, this Antora documentation is seamlessly merged with the pre-existing in-source documentation, which has been generated using various other tools.